+++ includes/cocart-core-functions.php
@@ -22,9 +22,9 @@ if ( ! defined( 'ABSPATH' ) ) {
  * @param string $name  Constant name.
  * @param mixed  $value Value.
  */
-function cocart_maybe_define_constant( $name, $value ) {
+function cocart_maybe_define_constant( $name, $value ) { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	if ( ! defined( $name ) ) {
-		define( $name, $value );
+		define( $name, $value ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.VariableConstantNameFound
 	}
 } // END cocart_maybe_define_constant()
 
@@ -40,7 +40,7 @@ function cocart_maybe_define_constant( $name, $value ) {
  *
  * @return string|null The timestamp the cart was created or expired, or null if invalid type.
  */
-function cocart_get_timestamp( $cart_key, $timestamp_type = 'created' ) {
+function cocart_get_timestamp( $cart_key, $timestamp_type = 'created' ) { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( 'created' === $timestamp_type ) {
@@ -74,7 +74,7 @@ function cocart_get_timestamp( $cart_key, $timestamp_type = 'created' ) {
  *
  * @return string
  */
-function cocart_get_source( $cart_key ) {
+function cocart_get_source( $cart_key ) { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	$value = $wpdb->get_var( // phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
@@ -94,7 +94,7 @@ function cocart_get_source( $cart_key ) {
  *
  * @return boolean Returns true or false if the session table exists.
  */
-function cocart_maybe_show_results() {
+function cocart_maybe_show_results() { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( $wpdb->get_var( "SHOW TABLES LIKE '{$wpdb->prefix}cocart_carts';" ) ) { // phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
@@ -115,7 +115,7 @@ function cocart_maybe_show_results() {
  *
  * @return int Number of carts in session.
  */
-function cocart_carts_in_session( $session = '' ) {
+function cocart_carts_in_session( $session = '' ) { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( ! cocart_maybe_show_results() ) {
@@ -150,7 +150,7 @@ function cocart_carts_in_session( $session = '' ) {
  *
  * @return int Number of carts expiring.
  */
-function cocart_count_carts_expiring() {
+function cocart_count_carts_expiring() { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( ! cocart_maybe_show_results() ) {
@@ -181,7 +181,7 @@ function cocart_count_carts_expiring() {
  *
  * @return int Number of carts active.
  */
-function cocart_count_carts_active() {
+function cocart_count_carts_active() { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( ! cocart_maybe_show_results() ) {
@@ -211,7 +211,7 @@ function cocart_count_carts_active() {
  *
  * @return int Number of carts expired.
  */
-function cocart_count_carts_expired() {
+function cocart_count_carts_expired() { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( ! cocart_maybe_show_results() ) {
@@ -241,7 +241,7 @@ function cocart_count_carts_expired() {
  *
  * @return int Number of carts created via the web.
  */
-function cocart_carts_source_web() {
+function cocart_carts_source_web() { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( ! cocart_maybe_show_results() ) {
@@ -271,7 +271,7 @@ function cocart_carts_source_web() {
  *
  * @return int Number of carts created via CoCart API.
  */
-function cocart_carts_source_headless() {
+function cocart_carts_source_headless() { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( ! cocart_maybe_show_results() ) {
@@ -301,7 +301,7 @@ function cocart_carts_source_headless() {
  *
  * @return int Number of carts created via other or unknown.
  */
-function cocart_carts_source_other() {
+function cocart_carts_source_other() { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	global $wpdb;
 
 	if ( ! cocart_maybe_show_results() ) {
@@ -328,7 +328,7 @@ function cocart_carts_source_other() {
  *
  * @since 4.2.0 Introduced.
  */
-function cocart_nocache_headers() {
+function cocart_nocache_headers() { // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedFunctionFound
 	cocart_maybe_define_constant( 'DONOTCACHEPAGE', true );
 	cocart_maybe_define_constant( 'DONOTCACHEOBJECT', true );
 	cocart_maybe_define_constant( 'DONOTCACHEDB', true );